icons: disable Shadow DOM inversion by default
This setting is specific to Chromium-based browsers.

Staring in GitLab 13.0, fill color for status icons will be using a CSS
variable that we can change without relying on hacks.

This makes it so that we don't need this option any longer, but it will
stay disabled so that users stuck on older versions of GitLab can enable
the option for status icons to look good.