icons: improve colors for CI status icons
This commit improves upon the last commit for CI status icons.

In order to keep maximum compatability with older instances of GitLab,
we're going back to old invert hack. Inverted colors will not exactly
match theme colors 1:1, but the icons will be dark on all instances.

If you're not using instances prior to GitLab 13.0 release (or newer, if
you're reading this in the future), you should disable this setting so
that it uses theme colors by removing invert hacks.

Closes #122.