labels: attempt to rewrite all project labels
This is going to be one hell of a task because labels evolved a lot over
the last two years that Dark-GitLab existed. The current implementation
has been tweaked on-and-off, and not even I can fully remember what's
going on in that entire block that's been commented out for now.

The main goal is to be be fully compatible with previous versions of
GitLab; the same way that other refactors and rewrites are, so that
Dark-GitLab can be usable on instances that are not up to date.

I doubt the solution is going to be "perfect" for everyone, but I hope
it will be just as good, or perhaps even better, than previous solution.