project: hack around empty space for languages bar
This commit will add a hack that fixes the last progress bar not having
rounded corners that has a width of a removed margin.

Due to how GitLab renders used languages, if there are too many of them,
they don't end up being rendered, and so an empty area is left behind.

References #149.