feat: split options for adjusting syntax colors
It is now possible to adjust saturation and lightness parts of HSL
format in separate options. This allows for more customizability than
before, which should be sufficient for most use-cases for now.

Ideally, we would want to have entirely separate syntax themes, or at
least "base" colors (like BG/FG/AC) for syntax colors. There's only so
much I can do on my own, so that task is going to my to-do list.

I also changed hue value of green color in order to match color-scheme
used in v1. It's not a 1:1 match, but it's "close enough" depending on
the values of aforementioned options.

RGB colors, a.k.a syntax colors, are got tweaked to use less saturation
and lightness, so that the default syntax colors aren't as bright as
in the previous release.

References #150.