Fixes color of OpenInYourIDE dropdown under clone
Under the `Clone` button on the project main page there is a `Open in
your IDE` section, under which is a `Visual Studio Code` hyperlink.
This hyperlink had the same dark foreground color as it's background.

This is because of the `.gl-new-dropdown-item .dropdown-item` rule had a
transparent background and #303030 foreground overwriting our own
color styles under `[class*="dropdown-menu"] li > a, [class*="dropdown-menu"]
li button, [class*="dropdown-menu"] li .menu-item`.

Since these elements `.dropdown-item`s have a transparent background
that means that tge actual background is dependent on the parent
element. Apparently there is a similar construct for the color CSS
property, namely the `inherit` CSS property value.