Fixes #171 white pipeline buttons
Replacing pipeline-graph CSS class with something similar. (this id
selector is not full width though!)

build CSS class is renamed to gl-build

ci-job-name-component does not directly contain gl-text-truncate

pipeline-tab-content is removed so the appropriate styling is removed
as well. This could be changed with a bit of trickery by using a
selector which is starting with js even though far from optimal and
using a `> div` CSS selector which is again terrible, but there are no
more specific selectors at that level. The only reasonable selector
would be [id^="pipeline-links-container-"] but that is too low in the
DOM to make the background lighter:(((

Signed-off-by: KartoffelCheetah <kartoffel.cheetah@gmail.com>